updates needed for python 3.12
had to dump python 3.9 because pymedphys doesn't support it in 0.41 and it doesn't support python 3.12 in 0.40
updated pythonapp.yml to use ubuntu 24.04 (not critical to python 3.12, but if we are going to move forward, jumping to a current
LTS of Ubuntu that has the required applications, like plastimatch, is part of it.
moved qt message handler to conftest.py
updated PySide6 version allowed, need latest for python 3.12.
updated dependencies as needed, and pinned shapely to the latest working version that didn't change the API (will need to update to 2.x at some point, not sure how much ripple that will have).