interrogation window
Added an interrogation window, works when clicking the windo button next to the fine tuning rotation and translation

added a none button next to it so the user can use the cutlines without moving the image or having a window. May need to look at disabling the cut lines for those buttons.

the 3 windows change to integration window when the button is pressed and dont change when the user changes teh Zoom